move csv related logic to its own Trait (task #2042)
